+++ js/check-positive-number.js
@@ -1,0 +1,1 @@
+export const isPositive = (number) => number >= 0;
